ver 0.3
json has a new look! :+1:
